+++ lib/src/upgrader.dart
@@ -1,11 +1,11 @@
 /*
- * Copyright (c) 2018-2022 Larry Aasen. All rights reserved.
+ * Copyright (c) 2018-2023 Larry Aasen. All rights reserved.
  */
 
 import 'dart:async';
+import 'dart:ui';
 
 import 'package:flutter/cupertino.dart';
-import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:http/http.dart' as http;
 import 'package:package_info_plus/package_info_plus.dart';
@@ -16,7 +16,7 @@ import 'package:version/version.dart';
 import 'appcast.dart';
 import 'itunes_search_api.dart';
 import 'play_store_search_api.dart';
-import 'upgrade_io.dart';
+import 'upgrade_os.dart';
 import 'upgrade_messages.dart';
 
 /// Signature of callbacks that have no arguments and return bool.
@@ -38,7 +38,7 @@ enum UpgradeDialogStyle { cupertino, material }
 
 /// A class to define the configuration for the appcast. The configuration
 /// contains two parts: a URL to the appcast, and a list of supported OS
-/// names, such as "android", "ios".
+/// names, such as "android", "fuchsia", "ios", "linux" "macos", "web", "windows".
 class AppcastConfiguration {
   final List<String>? supportedOS;
   final String? url;
@@ -107,12 +107,6 @@ class Upgrader {
   /// Return false when the default behavior should not execute.
   BoolCallback? onUpdate;
 
-  /// The [TargetPlatform] that identifies the platform on which the package is
-  /// currently executing. Defaults to [defaultTargetPlatform]. Note that
-  /// [TargetPlatform] does not include web, but includes mobile and desktop.
-  /// This parameter is normally used to change the target platform during testing.
-  final TargetPlatform platform;
-
   /// Called when the user taps outside of the dialog and [canDismissDialog]
   /// is false. Also called when the back button is pressed. Return true for
   /// the screen to be popped. Not used by [UpgradeCard].
@@ -137,8 +131,8 @@ class Upgrader {
   /// is logging metrics for your app.
   WillDisplayUpgradeCallback? willDisplayUpgrade;
 
-  /// The target operating system.
-  final String _operatingSystem = UpgradeIO.operatingSystem;
+  /// Provides information on which OS this code is running on.
+  final UpgraderOS upgraderOS;
 
   bool _displayed = false;
   bool _initCalled = false;
@@ -184,10 +178,10 @@ class Upgrader {
     this.minAppVersion,
     this.dialogStyle = UpgradeDialogStyle.material,
     this.cupertinoButtonTextStyle,
-    TargetPlatform? platform,
+    UpgraderOS? upgraderOS,
   })  : client = client ?? http.Client(),
         messages = messages ?? UpgraderMessages(),
-        platform = platform ?? defaultTargetPlatform {
+        upgraderOS = upgraderOS ?? UpgraderOS() {
     if (debugLogging) print("upgrader: instantiated.");
   }
 
@@ -235,18 +229,16 @@ class Upgrader {
 
       if (debugLogging) {
         print('upgrader: default operatingSystem: '
-            '${UpgradeIO.operatingSystem} ${UpgradeIO.operatingSystemVersion}');
-        print('upgrader: operatingSystem: $_operatingSystem');
-        print('upgrader: platform: $platform');
-        print('upgrader: defaultTargetPlatform: $defaultTargetPlatform');
+            '${upgraderOS.operatingSystem} ${upgraderOS.operatingSystemVersion}');
+        print('upgrader: operatingSystem: ${upgraderOS.operatingSystem}');
         print('upgrader: '
-            'isAndroid: ${UpgradeIO.isAndroid}, '
-            'isIOS: ${UpgradeIO.isIOS}, '
-            'isLinux: ${UpgradeIO.isLinux}, '
-            'isMacOS: ${UpgradeIO.isMacOS}, '
-            'isWindows: ${UpgradeIO.isWindows}, '
-            'isFuchsia: ${UpgradeIO.isFuchsia}, '
-            'isWeb: ${UpgradeIO.isWeb}');
+            'isAndroid: ${upgraderOS.isAndroid}, '
+            'isIOS: ${upgraderOS.isIOS}, '
+            'isLinux: ${upgraderOS.isLinux}, '
+            'isMacOS: ${upgraderOS.isMacOS}, '
+            'isWindows: ${upgraderOS.isWindows}, '
+            'isFuchsia: ${upgraderOS.isFuchsia}, '
+            'isWeb: ${upgraderOS.isWeb}');
       }
 
       if (_packageInfo == null) {
@@ -329,9 +321,9 @@ class Upgrader {
 
       // Get Android version from Google Play Store, or
       // get iOS version from iTunes Store.
-      if (platform == TargetPlatform.android) {
+      if (upgraderOS.isAndroid) {
         await _getAndroidStoreVersion(country: country, language: language);
-      } else if (platform == TargetPlatform.iOS) {
+      } else if (upgraderOS.isIOS) {
         final iTunes = ITunesSearchAPI();
         iTunes.debugEnabled = debugLogging;
         iTunes.client = client;
@@ -393,7 +385,8 @@ class Upgrader {
     // When there are no supported OSes listed, they are all supported.
     var supported = true;
     if (appcastConfig!.supportedOS != null) {
-      supported = appcastConfig!.supportedOS!.contains(_operatingSystem);
+      supported =
+          appcastConfig!.supportedOS!.contains(upgraderOS.operatingSystem);
     }
     return supported;
   }
@@ -820,7 +813,7 @@ class Upgrader {
     if (await canLaunchUrl(Uri.parse(_appStoreListingURL!))) {
       try {
         await launchUrl(Uri.parse(_appStoreListingURL!),
-            mode: UpgradeIO.isAndroid
+            mode: upgraderOS.isAndroid
                 ? LaunchMode.externalNonBrowserApplication
                 : LaunchMode.platformDefault);
       } catch (e) {